fix(suite): prevent cir. ESMs in TransactionReviewModal by moving common types and utils to separate files
