+++ packages/suite/src/components/suite/modals/ReduxModal/TransactionReviewModal/utils/index.ts
@@ -1,0 +1,13 @@
+import { SendState, StakeState } from '@suite-common/wallet-core';
+import { FormState } from '@suite-common/wallet-types';
+
+import { isStakeState } from '../types';
+
+export const hasTxValidityExpired = (deadline: number) => deadline <= Date.now();
+
+export const getTxType = (txInfoState: SendState | StakeState, precomposedForm: FormState) => {
+    const stakeType = isStakeState(txInfoState) ? 'stake' : undefined;
+    const tradeType = precomposedForm.trading?.activeSection ? 'trade' : undefined;
+
+    return stakeType ?? tradeType;
+};
